Support cached data, concept of "canonical" versions
A "canonical" version is one where the in-memory
state is exactly equal to what you get from
deserializing the chron response. Currently the
only non-canonical versions are those with cached
updates.

Updated get_conflicts to return a "canonical" bool.
This is not ideally-named because it is expected to
be true when the update is has conflicts. It really
represents "this update is non-canonical *solely*
because of extra stored data, and not because of
conflicts".